JHipster6105/src/test/javascript/spec/app/entities/agreement/agreement.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { AgreementService } from 'app/entities/agreement/agreement.service';
import { IAgreement, Agreement } from 'app/shared/model/agreement.model';

describe('Service Tests', () => {
  describe('Agreement Service', () => {
    let injector: TestBed;
    let service: AgreementService;
    let httpMock: HttpTestingController;
    let elemDefault: IAgreement;
    let expectedResult: IAgreement | IAgreement[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(AgreementService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Agreement(0, false, false, false, false, false);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Agreement', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Agreement()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Agreement', () => {
        const returnedFromService = Object.assign(
          {
            termsAgreement: true,
            gdprAgreement: true,
            photographyAgreement: true,
            marketingAgreement: true,
            medicalFitnessAgreement: true,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Agreement', () => {
        const returnedFromService = Object.assign(
          {
            termsAgreement: true,
            gdprAgreement: true,
            photographyAgreement: true,
            marketingAgreement: true,
            medicalFitnessAgreement: true,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Agreement', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
